Update the link BSS pointer during BSS reallocation on scan results
When updating the BSS during a scan results event, reallocation of the
BSS due to needing more room for IEs results in a new allocation and the
pointer changing. Update the link BSS pointer to the newly allocated BSS
similarly to the other cases that were covered previously. This is
needed to avoid use of freed memory in some MLO cases.

Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>

CRs-Fixed: 4054920
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 53303bb3e7437961d942f2b090c17d0b9f134e18
Change-Id: Ib4bb04e13a17b862c1156fe2121af263425f1f17